model_benchmarking/requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
tyro==0.9.35
tqdm==4.65.2
pillow==11.3.0
